refactor: Auth 또는 로그인 관련된 파일 이동
