+++ src/main/java/nextstep/auth/Authenticated.java
@@ -1,4 +1,4 @@
-package nextstep.member;
+package nextstep.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
